pyproject.toml
[project]
name = "docling-playground"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "docling>=2.24.0",
    "fastapi>=0.115.10",
    "ipykernel>=6.29.5",
    "lancedb>=0.20.0",
    "llama-index>=0.12.21",
    "llama-index-core>=0.12.21",
    "llama-index-embeddings-openai>=0.2.1",
    "llama-index-llms-openai>=0.3.25",
    "llama-index-vector-stores-lancedb>=0.3.0",
    "numpy>=1.26.4",
    "openai>=1.63.2",
    "pandas>=2.2.3",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "streamlit>=1.42.2",
    "tantivy==0.22.0",
    "tenacity>=9.0.0",
    "tiktoken>=0.9.0",
    "uvicorn>=0.34.0",
    "watchdog>=6.0.0",
]

[tool.uv]
no-build-isolation-package = ["torch", "torchvision", "torchaudio"]
